requirements.txt
beautifulsoup4==4.11.2
black==23.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cssselect==1.2.0
cymem==2.0.7
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
idna==3.4
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
murmurhash==1.0.9
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pathspec==0.11.0
pathy==0.10.1
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==3.1.0
preshed==3.0.8
pydantic==1.10.5
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
yarg==0.1.9
openapi
transformers
torch
rouge_score
gradio